requirements.txt
fastapi
uvicorn
pyyaml
requests
pydantic
pytest
httpx
